Permission base route

+++ src/todo/todo.service.ts
@@ -2,12 +2,13 @@ import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { PrismaService } from '../prisma/prisma.service';
+import { User } from '../interface/user.interface';
 
 @Injectable()
 export class TodoService {
   constructor(private prisma: PrismaService) {}
 
-  async create(createTodoDto: CreateTodoDto, user: any) {
+  async create(createTodoDto: CreateTodoDto, user: User) {
     try {
       createTodoDto.userId = user.id;
       return this.prisma.todo.create({ data: createTodoDto });
@@ -16,26 +17,32 @@ export class TodoService {
     }
   }
 
-  findAll() {
-    return `This action returns all todos`;
+  findAll(user: User) {
+    return this.prisma.todo.findMany({ where: { userId: user.id } });
   }
 
-  findOne(id: number) {
-    return `This action returns a #${id} todo`;
+  findOne(id: number, user: User) {
+    try {
+      return this.prisma.todo.findFirstOrThrow({
+        where: { id: id, userId: user.id },
+      });
+    } catch (e) {
+      throw new InternalServerErrorException(e);
+    }
   }
 
-  update(id: number, updateTodoDto: UpdateTodoDto) {
+  update(id: number, updateTodoDto: UpdateTodoDto, user: User) {
     try {
       return this.prisma.todo.update({
         data: updateTodoDto,
-        where: { id: id },
+        where: { id: id, userId: user.id },
       });
     } catch (error) {
       throw new InternalServerErrorException(error);
     }
   }
 
-  remove(id: number) {
-    return this.prisma.todo.delete({ where: { id: id } });
+  remove(id: number, user: User) {
+    return this.prisma.todo.delete({ where: { id: id, userId: user.id } });
   }
 }
